Major Changes all around. Implemented a new system of "expectation."
Basically, a formal definition now grows to represent everything within it. You can still modify, add, remove symbols, but this will also be automatically accomplished via the setting/adding of any component that UsesSymbols. This is to allow for the new dynamic in the interface where the modes will be distinguished by two general rules:
       1. Default mode = individual characters, no spaces needed (but can be used and will be removed/not represented)
       2. Custom mode = spaces designate symbol splitting. Symbols are automatically added to alphabets. There is no need for a parsing algorithm,  
       because the user will be responsible for designating what is and isnt a symbol. You will still be able to clik symbols in, but you will construct 
       the alphabet through building the language (i.e. productions or transitions/states). MODIFY will sitll be allowed. Applied via the uses symbols
       interface.