View has updated. Special Symbol addition/modification has been updated to reflect reactions as it should:
To SWITCH the StartVariable, one must simply set it to another symbol, this will not MODIFY the symbol that was previously there. To MODIFY the start symbol and have that change propagate throughout the code, you must modify the variable that is in the variable alphabet which corresponds to the start variable.